+++ unnamed_engine/src/renderer/viewport.rs
@@ -43,7 +43,7 @@ impl ViewportDesc {
       format,
       width: size.width,
       height: size.height,
-      present_mode: caps.present_modes[0],
+      present_mode: wgpu::PresentMode::Immediate,
       desired_maximum_frame_latency: 2,
       alpha_mode: caps.alpha_modes[0],
       view_formats: vec![],